+++ internal/transport/http/router.go
@@ -42,6 +42,8 @@ func RegisterRoutes(
 		api.PUT("/sites/:id", siteHandler.Update)
 		api.DELETE("/sites/:id", siteHandler.Delete)
 		api.POST("/sites/:id/deploy", siteHandler.Deploy)
+		api.GET("/sites/:id/ssl", siteHandler.GetSSLInfo)
+		api.POST("/sites/:id/ssl/renew", siteHandler.RenewSSL)
 		api.GET("/sites/:id/logs", siteHandler.GetLogs)
 		api.GET("/teams/access", teamHandler.List)
 		api.GET("/observability/ssl", observabilityHandler.SSL)